Corrected/improved the hit detection of items within PSTCollectionViewFlowLayout.layoutAttributesArray.
The existing code would return all items in the row tapped rather than just the item.  The new code now correctly adjusts the rectangle used to detect an intersection and only includes those items that intersect.

This seems (to me) to be more consistent with the behaviour of the Apple classes.