Handle the larger UICollectionView in iOS 8.
In addition to increasing padding in PSTCollectionView, this also changes PST class registration so that it does an all-or-nothing superclass replacement of the PSUI*_ classes. This prevents instances like the current case on iOS8-beta 5: PSUICollectionView_ has a superclass of PSTCollectionView, but the other types (e.g. PSUICollectionViewLayout_) had their superclasses replaced with UICollectionView*. This mixed mode is dangerous, as the PST* implementations rely on other PSUI*_ types maintaining their PST* superclasses.